Integrate ngram into gui
