--- ngram/.gitignore
@@ -1,3 +1,0 @@
-data
-model.txt
-
